+++ Modules/Microsoft365DSC/DSCResources/MSFT_EXOSafeLinksPolicy/MSFT_EXOSafeLinksPolicy.psm1
@@ -20,6 +20,7 @@ function Get-TargetResource
         [Boolean]
         $DeliverMessageAfterScan = $false,
 
+        #DEPRECATED
         [Parameter()]
         [Boolean]
         $DoNotAllowClickThrough = $true,
@@ -28,6 +29,7 @@ function Get-TargetResource
         [System.String[]]
         $DoNotRewriteUrls = @(),
 
+        #DEPRECATED
         [Parameter()]
         [Boolean]
         $DoNotTrackUserClicks = $true,
@@ -44,6 +46,7 @@ function Get-TargetResource
         [Boolean]
         $EnableSafeLinksForTeams = $false,
 
+        #DEPRECATED
         [Parameter()]
         [Boolean]
         $IsEnabled,
@@ -142,17 +145,20 @@ function Get-TargetResource
                 AdminDisplayName              = $SafeLinksPolicy.AdminDisplayName
                 CustomNotificationText        = $SafeLinksPolicy.CustomNotificationText
                 DeliverMessageAfterScan       = $SafeLinksPolicy.DeliverMessageAfterScan
-                DoNotAllowClickThrough        = $SafeLinksPolicy.DoNotAllowClickThrough
+                #Deprecated
+                #DoNotAllowClickThrough        = $SafeLinksPolicy.DoNotAllowClickThrough
                 DoNotRewriteUrls              = $SafeLinksPolicy.DoNotRewriteUrls
-                DoNotTrackUserClicks          = $SafeLinksPolicy.DoNotTrackUserClicks
+                #Deprecated
+                #DoNotTrackUserClicks          = $SafeLinksPolicy.DoNotTrackUserClicks
                 EnableForInternalSenders      = $SafeLinksPolicy.EnableForInternalSenders
                 EnableOrganizationBranding    = $SafeLinksPolicy.EnableOrganizationBranding
                 EnableSafeLinksForTeams       = $SafeLinksPolicy.EnableSafeLinksForTeams
-                IsEnabled                     = $SafeLinksPolicy.IsEnabled
+                #Deprecated
+                #IsEnabled                     = $SafeLinksPolicy.IsEnabled
                 ScanUrls                      = $SafeLinksPolicy.ScanUrls
                 UseTranslatedNotificationText = $SafeLinksPolicy.UseTranslatedNotificationText
                 Ensure                        = 'Present'
-                Credential            = $Credential
+                Credential                    = $Credential
                 ApplicationId                 = $ApplicationId
                 CertificateThumbprint         = $CertificateThumbprint
                 CertificatePath               = $CertificatePath
@@ -212,6 +218,7 @@ function Set-TargetResource
         [Boolean]
         $DeliverMessageAfterScan = $false,
 
+        #DEPRECATED
         [Parameter()]
         [Boolean]
         $DoNotAllowClickThrough = $true,
@@ -220,6 +227,7 @@ function Set-TargetResource
         [System.String[]]
         $DoNotRewriteUrls = @(),
 
+        #DEPRECATED
         [Parameter()]
         [Boolean]
         $DoNotTrackUserClicks = $true,
@@ -236,6 +244,7 @@ function Set-TargetResource
         [Boolean]
         $EnableSafeLinksForTeams = $false,
 
+        #DEPRECATED
         [Parameter()]
         [Boolean]
         $IsEnabled,
@@ -313,11 +322,31 @@ function Set-TargetResource
         }
         $SafeLinksPolicyParams.Remove('Identity') | Out-Null
         Write-Verbose -Message "Creating SafeLinksPolicy $($Identity)"
+
+        Write-Verbose -Message "Property DoNotTrackUserClicks is deprecated and will be ignored."
+        $SafeLinksPolicyParams.Remove("DoNotTrackUserClicks") | Out-Null
+
+        Write-Verbose -Message "Property DoNotAllowClickThrough is deprecated and will be ignored."
+        $SafeLinksPolicyParams.Remove("DoNotAllowClickThrough") | Out-Null
+
+        Write-Verbose -Message "Property IsEnabled is deprecated and will be ignored."
+        $SafeLinksPolicyParams.Remove("IsEnabled") | Out-Null
+
         New-SafeLinksPolicy @SafeLinksPolicyParams
     }
     elseif (('Present' -eq $Ensure ) -and ($null -ne $SafeLinksPolicy))
     {
         Write-Verbose -Message "Setting SafeLinksPolicy $($Identity) with values: $(Convert-M365DscHashtableToString -Hashtable $SafeLinksPolicyParams)"
+
+        Write-Verbose -Message "Property DoNotTrackUserClicks is deprecated and will be ignored."
+        $SafeLinksPolicyParams.Remove("DoNotTrackUserClicks") | Out-Null
+
+        Write-Verbose -Message "Property DoNotAllowClickThrough is deprecated and will be ignored."
+        $SafeLinksPolicyParams.Remove("DoNotAllowClickThrough") | Out-Null
+
+        Write-Verbose -Message "Property IsEnabled is deprecated and will be ignored."
+        $SafeLinksPolicyParams.Remove("IsEnabled") | Out-Null
+
         Set-SafeLinksPolicy @SafeLinksPolicyParams -Confirm:$false
     }
     elseif (('Absent' -eq $Ensure ) -and ($null -ne $SafeLinksPolicy))
@@ -349,6 +378,7 @@ function Test-TargetResource
         [Boolean]
         $DeliverMessageAfterScan = $false,
 
+        #DEPRECATED
         [Parameter()]
         [Boolean]
         $DoNotAllowClickThrough = $true,
@@ -357,6 +387,7 @@ function Test-TargetResource
         [System.String[]]
         $DoNotRewriteUrls = @(),
 
+        #DEPRECATED
         [Parameter()]
         [Boolean]
         $DoNotTrackUserClicks = $true,
@@ -373,6 +404,7 @@ function Test-TargetResource
         [Boolean]
         $EnableSafeLinksForTeams = $false,
 
+        #DEPRECATED
         [Parameter()]
         [Boolean]
         $IsEnabled,
@@ -442,6 +474,11 @@ function Test-TargetResource
     $ValuesToCheck.Remove('CertificateThumbprint') | Out-Null
     $ValuesToCheck.Remove('CertificatePath') | Out-Null
     $ValuesToCheck.Remove('CertificatePassword') | Out-Null
+
+    #DEPRECATED
+    $ValuesToCheck.Remove('DoNotAllowClickThrough') | Out-Null
+    $ValuesToCheck.Remove('DoNotTrackUserClicks') | Out-Null
+    $ValuesToCheck.Remove('IsEnabled') | Out-Null
 
     $TestResult = Test-M365DSCParameterState -CurrentValues $CurrentValues `
         -Source $($MyInvocation.MyCommand.Source) `