+++ Modules/Microsoft365DSC/DSCResources/MSFT_EXOSharedMailbox/MSFT_EXOSharedMailbox.psm1
@@ -13,8 +13,17 @@ function Get-TargetResource
         $PrimarySMTPAddress,
 
         [Parameter()]
+        [System.String]
+        $Alias,
+
+        # DEPRECATED
+        [Parameter()]
         [System.String[]]
         $Aliases,
+
+        [Parameter()]
+        [System.String[]]
+        $EmailAddresses,
 
         [Parameter()]
         [ValidateSet("Present", "Absent")]
@@ -71,6 +80,15 @@ function Get-TargetResource
     Add-M365DSCTelemetryEvent -Data $data
     #endregion
 
+    # Warning for deprecated parameter
+    if ($PSBoundParameters.ContainsKey("Aliases"))
+    {
+        Write-Warning "Aliases is deprecated. Please use EmailAddresses instead and remove Aliases from your configuration."
+        if ($null -eq $EmailAddresses) {
+            $EmailAddresses = $Aliases
+        }
+    }
+
     $nullReturn = $PSBoundParameters
     $nullReturn.Ensure = "Absent"
 
@@ -87,15 +105,15 @@ function Get-TargetResource
             return $nullReturn
         }
 
-        #region Email Aliases
-        $CurrentAliases = @()
+        #region EmailAddresses
+        $CurrentEmailAddresses = @()
 
         foreach ($email in $mailbox.EmailAddresses)
         {
             $emailValue = $email.Split(":")[1]
             if ($emailValue -and $emailValue -ne $mailbox.PrimarySMTPAddress)
             {
-                $CurrentAliases += $emailValue
+                $CurrentEmailAddresses += $emailValue
             }
         }
         #endregion
@@ -103,9 +121,10 @@ function Get-TargetResource
         $result = @{
             DisplayName           = $DisplayName
             PrimarySMTPAddress    = $mailbox.PrimarySMTPAddress.ToString()
-            Aliases               = $CurrentAliases
+            Alias                 = $mailbox.Alias
+            EmailAddresses        = $CurrentEmailAddresses
             Ensure                = "Present"
-            Credential    = $Credential
+            Credential            = $Credential
             ApplicationId         = $ApplicationId
             CertificateThumbprint = $CertificateThumbprint
             CertificatePath       = $CertificatePath
@@ -156,8 +175,17 @@ function Set-TargetResource
         $PrimarySMTPAddress,
 
         [Parameter()]
+        [System.String]
+        $Alias,
+
+        # DEPRECATED
+        [Parameter()]
         [System.String[]]
         $Aliases = @(),
+
+        [Parameter()]
+        [System.String[]]
+        $EmailAddresses = @(),
 
         [Parameter()]
         [ValidateSet("Present", "Absent")]
@@ -205,11 +233,11 @@ function Set-TargetResource
     $currentMailbox = Get-TargetResource @PSBoundParameters
 
     #region Validation
-    foreach ($alias in $Aliases)
+    foreach ($secondaryAlias in $EmailAddresses)
     {
-        if ($alias.ToLower() -eq $PrimarySMTPAddress.ToLower())
+        if ($secondaryAlias.ToLower() -eq $PrimarySMTPAddress.ToLower())
         {
-            throw "You cannot have the Aliases list contain the PrimarySMTPAddress"
+            throw "You cannot have the EmailAddresses list contain the PrimarySMTPAddress"
         }
     }
     #endregion
@@ -218,20 +246,39 @@ function Set-TargetResource
     $ConnectionMode = New-M365DSCConnection -Workload 'ExchangeOnline' `
         -InboundParameters $PSBoundParameters
 
+    # Warning for deprecated parameter
+    if ($PSBoundParameters.ContainsKey("Aliases"))
+    {
+        Write-Warning "Aliases is deprecated. Please use EmailAddresses instead and remove Aliases from your configuration."
+        if ($null -eq $EmailAddresses)
+        {
+            $EmailAddresses = $Aliases
+        }
+    }
+
     # CASE: Mailbox doesn't exist but should;
     if ($Ensure -eq "Present" -and $currentMailbox.Ensure -eq "Absent")
     {
         Write-Verbose -Message "Shared Mailbox '$($DisplayName)' does not exist but it should. Creating it."
         $emails = ""
-        foreach ($alias in $Aliases)
+        foreach ($secondaryAlias in $EmailAddresses)
         {
-            $emails += $alias + ","
+            $emails += $secondaryAlias + ","
         }
         $emails += $PrimarySMTPAddress
         $proxyAddresses = $emails -Split ','
-        $CurrentParameters.Aliases = $proxyAddresses
-        New-MailBox -Name $DisplayName -PrimarySMTPAddress $PrimarySMTPAddress -Shared:$true
-        Set-Mailbox -Identity $DisplayName -EmailAddresses @{add = $Aliases }
+        $CurrentParameters.EmailAddresses = $proxyAddresses
+        $NewMailBoxParameters = @{
+            Name                  = $DisplayName
+            PrimarySMTPAddress    = $PrimarySMTPAddress
+            Shared                = $true
+        }
+        if ($Alias)
+        {
+            $NewMailBoxParameters.Add("Alias", $Alias)
+        }
+        New-MailBox @NewMailBoxParameters
+        Set-Mailbox -Identity $DisplayName -EmailAddresses @{add = $EmailAddresses }
     }
     # CASE: Mailbox exists but it shouldn't;
     elseif ($Ensure -eq "Absent" -and $currentMailbox.Ensure -eq "Present")
@@ -242,45 +289,53 @@ function Set-TargetResource
     # CASE: Mailbox exists and it should, but has different values than the desired ones
     elseif ($Ensure -eq "Present" -and $currentMailbox.Ensure -eq "Present")
     {
-        # CASE: Email Aliases need to be updated
+        # CASE: EmailAddresses need to be updated
         Write-Verbose -Message "Shared Mailbox '$($DisplayName)' already exists, but needs updating."
-        $current = $currentMailbox.Aliases
-        $desired = $Aliases
+        $current = $currentMailbox.EmailAddresses
+        $desired = $EmailAddresses
         $diff = Compare-Object -ReferenceObject $current -DifferenceObject $desired
         if ($diff)
         {
-            # Add Aliases
-            Write-Verbose -Message "Updating the list of Aliases for the Shared Mailbox '$($DisplayName)'"
+            # Add EmailAddresses
+            Write-Verbose -Message "Updating the list of EmailAddresses for the Shared Mailbox '$($DisplayName)'"
             $emails = ""
-            $aliasesToAdd = $diff | Where-Object -FilterScript { $_.SideIndicator -eq '=>' }
-            if ($null -ne $aliasesToAdd)
+            $emailAddressesToAdd = $diff | Where-Object -FilterScript { $_.SideIndicator -eq '=>' }
+            if ($null -ne $emailAddressesToAdd)
             {
                 $emailsToAdd = ''
-                foreach ($alias in $aliasesToAdd)
+                foreach ($secondaryAlias in $emailAddressesToAdd)
                 {
-                    $emailsToAdd += $alias.InputObject + ","
+                    $emailsToAdd += $secondaryAlias.InputObject + ","
                 }
                 $emailsToAdd += $PrimarySMTPAddress
                 $proxyAddresses = $emailsToAdd -Split ','
 
-                Write-Verbose -Message "Adding the following email aliases: $emailsToAdd"
+                Write-Verbose -Message "Adding the following EmailAddresses: $emailsToAdd"
                 Set-Mailbox -Identity $DisplayName -EmailAddresses @{add = $proxyAddresses }
             }
-            # Remove Aliases
-            $aliasesToRemove = $diff | Where-Object -FilterScript { $_.SideIndicator -eq '<=' }
-            if ($null -ne $aliasesToRemove)
+            # Remove EmailAddresses
+            $emailAddressesToRemove = $diff | Where-Object -FilterScript { $_.SideIndicator -eq '<=' }
+            if ($null -ne $emailAddressesToRemove)
             {
                 $emailsToRemoved = ''
-                foreach ($alias in $aliasesToRemove)
+                foreach ($secondaryAlias in $emailAddressesToRemove)
                 {
-                    $emailsToRemoved += $alias.InputObject + ","
+                    $emailsToRemoved += $secondaryAlias.InputObject + ","
                 }
                 $emailsToRemoved += $PrimarySMTPAddress
                 $proxyAddresses = $emailsToRemoved -Split ','
 
-                Write-Verbose -Message "Removing the following email aliases: $emailsToRemoved"
+                Write-Verbose -Message "Removing the following EmailAddresses: $emailsToRemoved"
                 Set-Mailbox -Identity $DisplayName -EmailAddresses @{remove = $proxyAddresses }
             }
+        }
+        $current = $currentMailbox.Alias
+        $desired = $Alias
+        $diff = Compare-Object -ReferenceObject $current -DifferenceObject $desired
+        if ($diff)
+        {
+            Write-Verbose -Message "Updating Alias for the Shared Mailbox '$($DisplayName)'"
+            Set-Mailbox -Identity $DisplayName -Alias $Alias
         }
     }
 }
@@ -300,8 +355,17 @@ function Test-TargetResource
         $PrimarySMTPAddress,
 
         [Parameter()]
+        [System.String]
+        $Alias,
+
+        # DEPRECATED
+        [Parameter()]
         [System.String[]]
         $Aliases,
+
+        [Parameter()]
+        [System.String[]]
+        $EmailAddresses,
 
         [Parameter()]
         [ValidateSet("Present", "Absent")]
@@ -344,6 +408,16 @@ function Test-TargetResource
     Add-M365DSCTelemetryEvent -Data $data
     #endregion
 
+    # Warning for deprecated parameter
+    if ($PSBoundParameters.ContainsKey("Aliases"))
+    {
+        Write-Warning "Aliases is deprecated. Please use EmailAddresses instead and remove Aliases from your configuration."
+        if ($null -eq $EmailAddresses)
+        {
+            $EmailAddresses = $Aliases
+        }
+    }
+
     Write-Verbose -Message "Testing configuration of Office 365 Shared Mailbox $DisplayName"
 
     $CurrentValues = Get-TargetResource @PSBoundParameters
@@ -356,8 +430,9 @@ function Test-TargetResource
         -DesiredValues $PSBoundParameters `
         -ValuesToCheck @("Ensure", `
             "DisplayName", `
+            "Alias", `
             "PrimarySMTPAddress",
-        "Aliases")
+        "EmailAddresses")
 
     Write-Verbose -Message "Test-TargetResource returned $TestResult"
 
@@ -434,6 +509,7 @@ function Export-TargetResource
                 $params = @{
                     Credential    = $Credential
                     DisplayName           = $mailboxName
+                    Alias                 = $mailbox.Alias
                     ApplicationId         = $ApplicationId
                     TenantId              = $TenantId
                     CertificateThumbprint = $CertificateThumbprint