+++ CHANGELOG.md
@@ -1,5 +1,76 @@
 # Change log for Microsoft365DSC
 
+# UNRELEASED
+
+* TeamsUpdateManagementPolicy
+  * Changed the format of the UpdateTimeOfDay parameter to not include date as part of an export.
+    FIXES [#2062](https://github.com/microsoft/Microsoft365DSC/issues/2062)
+
+# 1.22.713.1
+
+* AADAuthorizationPolicy
+  * Initial release.
+* AADConditionalAccessPolicy
+  * Fixed issue for Included and Excluded properties where the last instance couldn't be removed.
+    FIXES [#2058](https://github.com/microsoft/Microsoft365DSC/issues/2058) & [#2079](https://github.com/microsoft/Microsoft365DSC/issues/2079)
+* EXOATPPolicyForO365
+  * Deprecated properties AllowClickThrough, EnableSafeLinksForO365Clients & TrackClicks.
+* EXOAuthenticationPolicyAssignment
+  * Initial release.
+* EXOCASMailboxSettings
+  * New resource to configure Exchange Online CAS Mailbox settings.
+* EXOSafeLinksPolicy
+  * Deprecated properties DoNotAllowClickThrough, DoNotTrackUserClicks & IsEnabled.
+* IntuneAppProtectionPolicyiOS
+   * Fixed issue with creation a new policies where it was complaining about invalid minimum versions.
+   * Fixed issues where creating new policies threw an error complaining about an invalid duration format.
+     FIXES [#2019](https://github.com/microsoft/Microsoft365DSC/issues/2019)
+   * Added the CustomBrowserProtocol paramters.
+     FIXES [#2009](https://github.com/microsoft/Microsoft365DSC/issues/2009)
+* IntuneDeviceAndAppManagementAssignmentFilter
+  * Initial release.
+* SCComplianceTag
+  * Fixed issue where FilePlanProperty was not properly applied unless another child property was also modified.
+* SPOSharingSettings
+    * Updated code to remove None as valid value for DefaultLinkPermission. If value is set to None default to Edit.
+* DEPENDENCIES
+  * Updated Microsoft.PowerApps.Administration.PowerShell to version 2.0.150.
+  * Updated MSCloudLoginAssistant to version 1.0.87.
+* MISC
+  * Made the Compare-M365DSCConfigurations cmdlet public.
+
+# 1.22.706.1
+
+* AADSecurityDefaults
+  * Initial release.
+* TeamsFederationConfiguration
+  * Added support for the AllowedDomains and BlockedDomains properties.
+* TeamsVoiceRoutingPolicy
+  * Remove unsupported Confirm parameter from Remove-CsOnlineVoiceRoutingPolicy cmdlet (Confirm parameter is no longer available for MicrosoftTeams PowerShell module 4.4.1+).
+    FIXES #2055
+* DEPENDENCIES
+  * Updated MicrosoftTeams to version 4.5.0.
+  * Updated Pnp.PowerShell to version 1.11.0.
+
+# 1.22.629.1
+
+* EXOMalwareFilterPolicy
+  * Parameters Action, EnableExternalSenderNotifications and EnableInternalSenderNotifications are deprecated and will be removed in future. These parameters are no longer available in EXO, only in onprem Exchange. Please remove these parameters from your configuration.
+  FIXES #2025
+  * Added support for FileTypeAction parameter.
+* EXOSharedMailbox
+  * Fix using umlauts in displayname by allowing to set alias.
+    FIXES #1921
+  * Rename parameter Aliases to EmailAddresses. Aliases is now deprecated.
+* DEPENDENCIES
+  * Updated DSCParser to version 1.3.0.5
+  * Updated Microsoft.Graph.* modules to version 1.10.0.
+  * Updated MSCloudLoginAssistant to version 1.0.86.
+    Fixes two authentication issues: #2000 and #2007
+* MISC
+  * New Delta Report: removed mandatory restrictions on the OutputPath parameter.
+    FIXES #2029
+
 # 1.22.622.1
 
 * TeamsMessagingPolicy